drone_backend/requirements.txt
google-generativeai
python-dotenv
flask